+++ tests/meltano/core/test_schedule_service.py
@@ -2,6 +2,9 @@ from datetime import datetime
 from unittest import mock
 
 import pytest
+from meltano.core.plugin import PluginType
+from meltano.core.plugin.project_plugin import ProjectPlugin
+from meltano.core.project_plugins_service import PluginAlreadyAddedException
 from meltano.core.schedule_service import (
     Schedule,
     ScheduleAlreadyExistsError,
@@ -27,6 +30,17 @@ def create_schedule():
         return Schedule(name=name, **attrs)
 
     return make
+
+
+@pytest.fixture(scope="class")
+def custom_tap(project_add_service):
+    tap = ProjectPlugin(
+        PluginType.EXTRACTORS, name="tap-custom", namespace="tap_custom"
+    )
+    try:
+        return project_add_service.add_plugin(tap)
+    except PluginAlreadyAddedException as err:
+        return err.plugin
 
 
 class TestScheduleService:
@@ -150,3 +164,27 @@ class TestScheduleService:
                 ],
                 env={"TAP_MOCK_TEST": "overridden", "TAP_MOCK_SECURE": "overridden"},
             )
+
+    def test_find_namespace_schedule(
+        self, subject, tap, create_schedule, project_plugins_service
+    ):
+        schedule = create_schedule(tap.name)
+        subject.add_schedule(schedule)
+        with mock.patch(
+            "meltano.core.project_plugins_service.ProjectPluginsService",
+            return_value=project_plugins_service,
+        ):
+            found_schedule = subject.find_namespace_schedule(tap.namespace)
+            assert found_schedule.extractor == tap.name
+
+    def test_find_namespace_schedule_custom_extractor(
+        self, subject, create_schedule, custom_tap, project_plugins_service
+    ):
+        schedule = Schedule(name="tap-custom", extractor="tap-custom")
+        subject.add_schedule(schedule)
+        with mock.patch(
+            "meltano.core.project_plugins_service.ProjectPluginsService",
+            return_value=project_plugins_service,
+        ):
+            found_schedule = subject.find_namespace_schedule(custom_tap.namespace)
+            assert found_schedule.extractor == custom_tap.name